requirements.txt
streamlit
pytesseract
Pillow
PyMuPDF
chromadb
sentence-transformers
spacy
streamlit==1.31.0
requests==2.31.0
pandas==2.2.0
python-dotenv==1.0.0